mobile-224-PhotoPinch+ZoomScreenFixes: fixed iss ue with image not taking up more of the screen, fixed back button being cut off, improved pan and pinch to alivaite shuddering
