+++ compnents/reusables/animatedFullScreenModal.js
@@ -9,9 +9,10 @@ import { ActiveTutorialIDContext } from "../contexts/activeTutorialIDContext";
 import { FullScreenModalContext } from "../contexts/fullScreenModalContext";
 
 import OnboardingTest from "../tutorial/onboarding";
-import PhotoBoxModel from "../modals/photoBoxModal";
+import PhotoBoxModel from '../screens/photoBox/photoBoxModal';
 import CommentsModal from "../modals/commentsModal";
 import EditScreenParallax from '../screens/edits/editsParallax';
+import { colors } from "../styles";
 
 const windowWidth = Dimensions.get("window").width;
 const windowHeight = Dimensions.get("screen").height;
@@ -57,5 +58,6 @@ const styles = StyleSheet.create({
     width: windowWidth,
     zIndex: 50,
     left: 0,
+    backgroundColor: colors.themeBlack
   },
 });
